Use parallel (de)compressing for zstd archives
Use `pzstd` for parallel decompressing.

Use `--threads` for parallel `xz` compressing.  The number of threads used for
`xz` is 'let `xz` decide' (which defaults to 'number of processor(s) on the
system'), except for 32-bit platforms where 2 threads are used.  This is the
same as what `makepkg` uses for `xz` by default.